change HELP_TREE to const instead of enum.
