+++ .changeset/stale-lemons-hang.md
@@ -1,0 +1,5 @@
+---
+'@sap-ux/guided-answers-helper': patch
+---
+
+fix export of HELP_TREE and simplify structure
