+++ packages/guided-answers-helper/src/guided-answers-icon_svg_base64.ts
