--- packages/guided-answers-helper/src/guided-answers/index.ts
@@ -1,2 +1,0 @@
-export * from './guided-answers-icon_svg_base64';
-export * from './help-topics';
